Refactor:liked page parallel routing
